pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "ipl-predictor"
version = "0.1.0"
description = "IPL score and win-probability predictor built from Cricsheet ball-by-ball data."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "numpy>=1.26,<3",
  "pandas>=2.2,<3",
  "scikit-learn>=1.7,<1.8",
  "joblib>=1.4,<2",
  "Flask>=3.1,<4",
  "Flask-Login>=0.6,<1",
  "Flask-WTF>=1.2,<2",
  "Flask-Limiter>=3.8,<4",
  "redis>=5.0,<6",
  "SQLAlchemy>=2.0,<3",
  "alembic>=1.13,<2",
  "marshmallow>=3.22,<4",
  "email-validator>=2.2,<3",
  "psycopg2-binary>=2.9,<3",
  "streamlit>=1.44,<2",
  "xgboost>=2.1,<3",
  "catboost>=1.2,<2",
  "matplotlib>=3.9,<4",
  "python-docx>=1.1,<2",
  "pillow>=11.2,<12",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4,<9",
]
dl = [
  "torch>=2.7,<3",
]

[tool.setuptools]
packages = ["ipl_predictor"]
